Move projectionChartUtils to shared/ folder
